% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{cl_aspect_ratio}
\alias{cl_aspect_ratio}
\title{AspectRatio component}
\usage{
cl_aspect_ratio(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
AspectRatio component
}
\references{
\href{https://mantine.dev/core/aspect-ratio/}{AspectRatio documentation}
}
\seealso{
Other layout: 
\code{\link{cl_center}()},
\code{\link{cl_container}()},
\code{\link{cl_flex}()},
\code{\link{cl_grid}()},
\code{\link{cl_grid_col}()},
\code{\link{cl_group}()},
\code{\link{cl_simple_grid}()},
\code{\link{cl_space}()},
\code{\link{cl_stack}()}
}
\concept{layout}
