% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{cl_center}
\alias{cl_center}
\title{Center component}
\usage{
cl_center(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Center component
}
\references{
\href{https://mantine.dev/core/center/}{Center documentation}
}
\seealso{
Other layout: 
\code{\link{cl_aspect_ratio}()},
\code{\link{cl_container}()},
\code{\link{cl_flex}()},
\code{\link{cl_grid}()},
\code{\link{cl_grid_col}()},
\code{\link{cl_group}()},
\code{\link{cl_simple_grid}()},
\code{\link{cl_space}()},
\code{\link{cl_stack}()}
}
\concept{layout}
