% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{cl_container}
\alias{cl_container}
\title{Container component}
\usage{
cl_container(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Container component
}
\references{
\href{https://mantine.dev/core/container/}{Container documentation}
}
\seealso{
Other layout: 
\code{\link{cl_aspect_ratio}()},
\code{\link{cl_center}()},
\code{\link{cl_flex}()},
\code{\link{cl_grid}()},
\code{\link{cl_grid_col}()},
\code{\link{cl_group}()},
\code{\link{cl_simple_grid}()},
\code{\link{cl_space}()},
\code{\link{cl_stack}()}
}
\concept{layout}
