% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{cl_flex}
\alias{cl_flex}
\title{Flex component}
\usage{
cl_flex(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Flex component
}
\references{
\href{https://mantine.dev/core/flex/}{Flex documentation}
}
\seealso{
Other layout: 
\code{\link{cl_aspect_ratio}()},
\code{\link{cl_center}()},
\code{\link{cl_container}()},
\code{\link{cl_grid}()},
\code{\link{cl_grid_col}()},
\code{\link{cl_group}()},
\code{\link{cl_simple_grid}()},
\code{\link{cl_space}()},
\code{\link{cl_stack}()}
}
\concept{layout}
