% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{cl_grid_col}
\alias{cl_grid_col}
\title{Grid col component}
\usage{
cl_grid_col(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Grid col component
}
\references{
\href{https://mantine.dev/core/grid/}{Grid documentation}
}
\seealso{
Other layout: 
\code{\link{cl_aspect_ratio}()},
\code{\link{cl_center}()},
\code{\link{cl_container}()},
\code{\link{cl_flex}()},
\code{\link{cl_grid}()},
\code{\link{cl_group}()},
\code{\link{cl_simple_grid}()},
\code{\link{cl_space}()},
\code{\link{cl_stack}()}

Other grid: 
\code{\link{cl_grid}()}
}
\concept{grid}
\concept{layout}
