% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{cl_space}
\alias{cl_space}
\title{Space component}
\usage{
cl_space(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Space component
}
\references{
\href{https://mantine.dev/core/space/}{Space documentation}
}
\seealso{
Other layout: 
\code{\link{cl_aspect_ratio}()},
\code{\link{cl_center}()},
\code{\link{cl_container}()},
\code{\link{cl_flex}()},
\code{\link{cl_grid}()},
\code{\link{cl_grid_col}()},
\code{\link{cl_group}()},
\code{\link{cl_simple_grid}()},
\code{\link{cl_stack}()}
}
\concept{layout}
