% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantine_page.R
\name{clemantine_page}
\alias{clemantine_page}
\title{clemantine_page}
\usage{
clemantine_page(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
clemantine_page
}
\references{
\href{https://mantine.dev/theming/mantine-provider/}{MantineProvider documentation}
}
